# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dccba68f1ef4518a53834d0
  title: wishbox
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Category  -----
  /category/:
    post:
      x-skaffolder-id: 5dccbdbaf1ef4518a538358f
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dccbc38f1ef4518a5383562
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Category
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                desc:
                  type: string
                isPublic:
                  type: boolean
                name:
                  type: string
    get:
      x-skaffolder-id: 5dccbdbbf1ef4518a5383597
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dccbc38f1ef4518a5383562
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Category
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /category/{id}:
    delete:
      x-skaffolder-id: 5dccbdbbf1ef4518a5383598
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dccbc38f1ef4518a5383562
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dccbdbaf1ef4518a5383591
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dccbc38f1ef4518a5383562
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    post:
      x-skaffolder-id: 5dccbdbaf1ef4518a5383590
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dccbc38f1ef4518a5383562
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
  /category/findByowner/{key}:
    get:
      x-skaffolder-id: 5dccbdb3f1ef4518a538357f
      x-skaffolder-name: findByowner
      x-skaffolder-id-resource: 5dccbc38f1ef4518a5383562
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: findByowner
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByowner
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByowner/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByowner
      tags:
        - Category
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa owner da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    
  #-----  Item  -----
...
  /user/:
    post:
      x-skaffolder-id: 5dccba69f1ef4518a53834df
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dccba69f1ef4518a53834da
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dccba69f1ef4518a53834e3
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dccba69f1ef4518a53834da
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dccba69f1ef4518a53834e4
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dccba69f1ef4518a53834da
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dccba69f1ef4518a53834e0
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dccba69f1ef4518a53834da
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dccba69f1ef4518a53834de
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dccba69f1ef4518a53834da
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
    Category:
      x-skaffolder-id: 5dccbc38f1ef4518a5383562
      x-skaffolder-id-db: 5dccba68f1ef4518a53834d1
      x-skaffolder-id-entity: 5dccbc38f1ef4518a538355e
      x-skaffolder-url: /category
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        desc:
          type: string
          x-skaffolder-id-attr: 5dccbc38f1ef4518a5383560
          x-skaffolder-type: String
        isPublic:
          type: boolean
          x-skaffolder-id-attr: 5dccbc38f1ef4518a5383561
          x-skaffolder-type: Boolean
        name:
          type: string
          x-skaffolder-id-attr: 5dccbc38f1ef4518a538355f
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
        owner:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dccbc38f1ef4518a538355e
          x-skaffolder-ent2: 5dccba68f1ef4518a53834d2
      required:
        - _id
        - name
...
    User:
      x-skaffolder-id: 5dccba69f1ef4518a53834da
      x-skaffolder-id-db: 5dccba68f1ef4518a53834d1
      x-skaffolder-id-entity: 5dccba68f1ef4518a53834d2
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5dccba68f1ef4518a53834d5
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dccba68f1ef4518a53834d6
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dccba68f1ef4518a53834d4
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dccba68f1ef4518a53834d8
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dccba68f1ef4518a53834d7
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dccba68f1ef4518a53834d3
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    
  # Databases
...